Fix wrong hyperlink targets for listings
```caption``` automatically loads ```hypcap``` with correct settings for
moving a hyperlink target to the begin of the surrounding floating environment.
When ```hypcap``` has been loaded manually before, the target of links
to minted-provided listings is placed at the caption, which is below the
listing.

While the ```hypcap``` option is a default setting for ```caption```,
setting it explicitly helps documentation.